Create README.md
GADSLeaderBoard is a practice project given to Associate Android Developers scholars for GADS 2020.
This application implements a mobile design of the GADS 2020 LeaderBoard. It is aimed to display Top 20 learners in the learning leaders and Top 20 ;earners in the skill IQ leaders category. 
This app is aimed to help us understand the concept of Multi-Activity project, ViewPager, Customizing Toolbar, RecyclerView, making a network request through Retrofit and working with google forms et al.

All rights reserved.